backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pydantic==2.4.2
pokerkit

# Development dependencies (optional, include if needed for dev environments)
pytest==7.4.3
httpx==0.25.0
pytest-asyncio==0.21.1
